+++ TagsCloudVisualization/TagsCloudVisualization/Spiral.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Drawing;
